+++ packages/cs-shared/UBDomain.js
@@ -121,6 +121,17 @@ function UBDomain (domainInfo) {
     me.orderedModels.push(me.models[modelCode])
   })
   me.orderedModels.sort((a, b) => a.order - b.order)
+
+  /**
+   * Array of vendor models names
+   * @type {Array<string>}
+   */
+  this.vendorModels = domainInfo.vendorModels ? domainInfo.vendorModels.split(':') : []
+  /**
+   * Array of customer models names
+   * @type {Array<string>}
+   */
+  this.customerModels = domainInfo.customerModels ? domainInfo.customerModels.split(':') : []
 }
 
 /**