+++ packages/base/base.js
@@ -8,7 +8,7 @@ const FileBasedStoreLoader = require('./FileBasedStoreLoader')
 const Worker = require('./worker')
 // eslint-disable-next-line camelcase
 const uba_common = require('./uba_common')
-const createDBConnectionPool = require('./DBConnections')
+const { createDBConnectionPool, releaseDBConnectionPool } = require('./DBConnections')
 /**
  * Classes, common for CLI and server side
  * @module @unitybase/base
@@ -72,6 +72,10 @@ module.exports = {
    * Create a database connection pool from current config
    */
   createDBConnectionPool,
+  /**
+   * Release previously created connection pool
+   */
+  releaseDBConnectionPool,
   /**
    * Allows to define a tokenized string and pass an arbitrary number of arguments to replace the tokens.  Each
    * token must be unique, and must increment in the format {0}, {1}, etc.