+++ packages/uba/_initialData/ActiveDirectryImport/installAD.ps1
@@ -1,2 +1,2 @@
 import-module servermanager
-Add-WindowsFeature -Name "RSAT-AD-PowerShell" –IncludeAllSubFeature
+Add-WindowsFeature -Name "RSAT-AD-PowerShell" -IncludeAllSubFeature