+++ packages/cs-shared/CHANGELOG.md
@@ -6,14 +6,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
- - new properties `UBDomain.customerModels` & `UBDomain.vendorModels` - array of customer/vendor models names.
-   Require UB server to be >= 5.18.17. For earlier versions returns empty arrays. 
 
 ### Changed
 
 ### Deprecated
 
 ### Removed
+
+### Fixed
+
+## [5.5.8] - 2020-11-05
+### Added
+ - new properties `UBDomain.customerModels` & `UBDomain.vendorModels` - array of customer/vendor models names.
+   Require UB server to be >= 5.18.17. For earlier versions returns empty arrays. 
 
 ### Fixed
  - DBConnectionConfig type definition: added missing `isDefault` property